requirements.txt
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
authlib==1.6.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
cyclopts==3.22.5
distro==1.9.0
dnspython==2.7.0
docstring-parser==0.17.0
docutils==0.22
email-validator==2.2.0
exceptiongroup==1.3.0
fastmcp==2.11.2
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
isodate==0.7.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
langchain-core==0.3.74
langchain-groq==0.3.7
langchain-mcp-adapters==0.1.9
langgraph==0.6.4
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.0
langsmith==0.4.13
lazy-object-proxy==1.11.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mcp==1.12.4
mdurl==0.1.2
more-itertools==10.7.0
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
parse==1.20.2
pathable==0.4.4
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pyperclip==1.9.0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.27.0
six==1.17.0
sniffio==1.3.1
sse-starlette==3.0.2
starlette==0.47.2
tenacity==9.1.2
typing-extensions==4.14.1
typing-inspection==0.4.1
urllib3==2.5.0
uvicorn==0.35.0
werkzeug==3.1.1
xxhash==3.5.0
zstandard==0.23.0
